Created essence genre, added feature search movies by genre

+++ database/migrations/2024_01_04_155644_add_genre_id_column_to_movies_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('movies', function (Blueprint $table) {
-            $table->unsignedTinyInteger('genre_id');
+            $table->integer('genre_id');
         });
     }
